Add deprecated property to IApiDescription - #3138
