+++ package.json
@@ -54,6 +54,7 @@
     "markdown-component-loader": "^1.1.0",
     "markdown-it-anchor": "^5.0.2",
     "mini-css-extract-plugin": "0.4.3",
+    "moment": "^2.24.0",
     "node-sass": "^4.12.0",
     "node-sass-tilde-importer": "^1.0.2",
     "npm-run-all": "^4.1.3",
