+++ src/apiDefs/benefits.ts
@@ -1,4 +1,4 @@
-import { IApiDescription } from '.';
+import { IApiDescription } from './schema';
 
 const swaggerHost : string = process.env.REACT_APP_VETSGOV_SWAGGER_API!;
 const benefitsApis : IApiDescription[] = [
